+++ src/SimpleResize_AVX2.cpp
@@ -48,8 +48,8 @@ static FORCE_INLINE void simpleResize_uint8_t_horizontal_8px_avx2(const SimpleRe
 
 // Thread-safe.
 void simpleResize_uint8_t_avx2(const SimpleResize *simple,
-                               uint8_t *dstp, int dst_stride,
-                               const uint8_t *srcp, int src_stride,
+                               uint8_t *dstp, ptrdiff_t dst_stride,
+                               const uint8_t *srcp, ptrdiff_t src_stride,
                                int horizontal_vectors) {
     (void)horizontal_vectors;
 
@@ -136,8 +136,8 @@ static FORCE_INLINE void simpleResize_int16_t_horizontal_8px_avx2(const SimpleRe
 
 // Thread-safe.
 void simpleResize_int16_t_avx2(const SimpleResize *simple,
-                               int16_t *dstp, int dst_stride,
-                               const int16_t *srcp, int src_stride,
+                               int16_t *dstp, ptrdiff_t dst_stride,
+                               const int16_t *srcp, ptrdiff_t src_stride,
                                int horizontal_vectors) {
     int16_t *workp = (int16_t *)malloc(simple->src_width * sizeof(int16_t));
